+++ src/router/index.js
@@ -6,7 +6,6 @@ import AppointmentRoute from "@/router/Appointment";
 import MyClubsPage from "@/pages/MyClubsPage.vue";
 import MyAppointmentsPage from "@/pages/promise/MyAppointmentsPage.vue";
 import JoinedClubsPage from "@/pages/club/JoinedClubsPage.vue";
-import JoinedAppointmentsPage from "@/pages/promise/JoinedAppointmentsPage.vue";
 
 const router = createRouter({
   history: createWebHistory(),
@@ -22,7 +21,6 @@ const router = createRouter({
     },
     // (임시) 아래 두 줄을 다른 라우터 파일에 분리?
     { path: "/joinedClubs", name: "JoinedClubs", component: JoinedClubsPage },
-    { path: "/joinedAppointments", name: "JoinedAppointments", component: JoinedAppointmentsPage },
 
     ...AppointmentRoute,
     // { path: '/input', name: "Input", component: Input },
